django-starter/requirements.txt
django>=4.0.2
ipython>=8.0.1
psycopg2>=2.9.3
djangorestframework>=3.13.1